requirements.txt
# ============================
# Core
# ============================
pandas>=2.0
numpy>=1.24
python-dotenv>=1.0
pyyaml>=6.0

# ============================
# Visualización
# ============================
matplotlib>=3.7

# ============================
# Google Cloud Platform
# ============================
google-cloud-aiplatform>=1.38
google-cloud-texttospeech>=2.14
google-cloud-storage>=2.10
google-genai>=1.0

# ============================
# Telegram
# ============================
python-telegram-bot>=21.0

# ============================
# Reportes PDF
# ============================
fpdf2>=2.7

# ============================
# Testing
# ============================
pytest>=7.0
